+++ CHANGELOG.md
@@ -6,7 +6,11 @@
 
 ##### Enhancements
 
-* None.
+* `TrailingCommaRule` now only triggers when a declaration is multi-line
+  when using `mandatory_comma: true`.  
+  [Marcelo Fabri](https://github.com/marcelofabri)
+  [#910](https://github.com/realm/SwiftLint/issues/910)
+  [#911](https://github.com/realm/SwiftLint/issues/911)
 
 ##### Bug Fixes