+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Master
+## 0.13.2: Light Cycle
 
 ##### Breaking